% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframe.R
\name{as_iframe}
\alias{as_iframe}
\title{Embed bsselect dropdowns in Rmd documents as iframes}
\usage{
as_iframe(widget, file = NULL, height = "500", width = "100\%")
}
\arguments{
\item{widget}{The bsselect htmlwidget object you want to embed}

\item{file}{The name of the file you'd like to save to}

\item{height}{The height of the iframe}

\item{width}{The width of the iframe}
}
\description{
Embed bsselect dropdowns in Rmd documents as iframes
}

