% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsselect.R
\name{bsselect}
\alias{bsselect}
\title{bsselect}
\usage{
bsselect(vector, selected = NULL, type = c("text", "img", "iframe"),
  frame_height = "500", frame_width = "100\%", align_right = FALSE,
  dropup_auto = TRUE, header = FALSE, live_search = FALSE,
  box_width = FALSE, live_search_style = "contains", show_tick = FALSE,
  size = "auto", style = NULL, width = NULL, height = NULL,
  elementId = NULL, ...)
}
\arguments{
\item{vector}{A named vector of values to send to the bootstrap-select dropdown menu.}

\item{selected}{The selected option; currently disabled}

\item{type}{One of \code{"text"}, \code{"img"}, or \code{"iframe"}.}

\item{frame_height}{The height of the image or iframe.  Defaults to 500px.}

\item{frame_width}{The width of the image or iframe.}

\item{align_right}{Whether to align the menu to the right instead of the left.  Defaults to FALSE.}

\item{dropup_auto}{Creates a drop-up menu automatically if there is not enough space to drop down.  Defaults to TRUE.}

\item{header}{Adds a header to the top of the menu.}

\item{live_search}{When TRUE, adds a search box to the drop-down menu.  Defaults to FALSE.}

\item{box_width}{One of "auto", "fit", a css-width in pixels, or FALSE.}

\item{live_search_style}{One of "contains" (the default) or "startsWith".}

\item{show_tick}{Whether to show a check mark next to the selected option.  Defaults to FALSE.}

\item{size}{The number of items to show in the window.  Defaults to "auto"; if FALSE, shows every option.}

\item{style}{Add the value to the button's style.  Options include "btn-primary", "btn-info", "btn-success", "btn-warning", and "btn-danger".}

\item{width}{The width of the htmlwidget.}

\item{height}{The height of the htmlwidget.}

\item{elementId}{The element ID.}
}
\description{
Function to generate a drop-down menu with the bootstrap-select jQuery plugin in an R Markdown document
}
\seealso{
\url{https://silviomoreto.github.io/bootstrap-select/}
}

