% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsselectR_package.R
\docType{package}
\name{bsselectR}
\alias{bsselectR}
\alias{bsselectR-package}
\title{Add bootstrap-select dropdown menus to R Markdown documents}
\description{
This R package enables users to display text, images, or iframes from dropdown menus in R Markdown documents, which are rendered with the bootstrap-select jQuery plugin.
}
\note{
bootstrap-select authors: Silvio Moreto, Ana Carolina, caseyjhol, Matt Bryson, and t0xicCode.
}
\author{
Kyle Walker
}
\seealso{
\url{https://silviomoreto.github.io/bootstrap-select/}
}

